requirements.txt
langchain-core==0.3.28 
langchain==0.3.13
langchain_community
langgraph==0.2.60  
langsmith==0.2.4 
google-genai==0.3.0
IPython
psycopg2-binary==2.9.9
pydantic-core==2.23.3  
pydantic-settings==2.5.2  
pydantic==2.9.1
python-dotenv==1.0.1
sqlalchemy==2.0.34  
sqlmodel==0.0.22
uvicorn==0.25.0
fastapi-cli[standard]==0.0.5  
fastapi==0.115.3 
pyttsx3
SpeechRecognition
setuptools
beautifulsoup4
tiktoken
google-generativeai